tests/wpt/web-platform-tests/bluetooth/idl-Bluetooth.html

<!DOCTYPE html>
<title>Bluetooth interface</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script>
'use strict';

test(() => {
  assert_throws(new TypeError(), () => new Bluetooth(),
                'the constructor should not be callable with "new"');
  assert_throws(new TypeError(), () => Bluetooth(),
                'the constructor should not be callable');

  // Bluetooth implements BluetoothDiscovery;
  assert_true('requestDevice' in navigator.bluetooth);
  assert_equals(navigator.bluetooth.requestDevice.length, 0);
}, 'Bluetooth IDL test');

</script>
